refactor: Moved the OAuth2LoginSuccessHandler file to the security package.
